separate checkboxes into separate components
