+++ packages/table/src/useLeafyGreenTable/TableHeaderCheckbox.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Row, Table } from '@tanstack/react-table';
+import { Table } from '@tanstack/react-table';
 
 import Checkbox from '@leafygreen-ui/checkbox';
 import { cx } from '@leafygreen-ui/emotion';
@@ -27,32 +27,6 @@ export const TableHeaderCheckbox = ({
       indeterminate={table.getIsSomeRowsSelected()}
       onChange={table.getToggleAllRowsSelectedHandler()}
       aria-label="Select all rows"
-    />
-  );
-};
-
-export const TableCellCheckbox = ({
-  row,
-  table,
-}: {
-  table: Table<LGTableDataType<LGRowData>>;
-  row: Row<LGTableDataType<LGRowData>>;
-}) => {
-  const { theme } = useDarkMode();
-  const { disabled: rowIsDisabled } = useRowContext();
-  return (
-    <Checkbox
-      className={cx({
-        [disabledTableRowCheckStyles[theme]]: rowIsDisabled,
-      })}
-      disabled={rowIsDisabled}
-      checked={row.getIsSelected()}
-      indeterminate={row.getIsSomeSelected()}
-      onChange={row.getToggleSelectedHandler()}
-      aria-label={`Select row ${row.id}`}
-      aria-controls={`lg-table-row-${row.id}`}
-      // Don't animate if _all_ rows have been checked (usually, if header row is clicked). Not the _best_ check, but it mostly works
-      animate={!table.getIsAllRowsSelected()}
     />
   );
 };
