+++ packages/table/src/useLeafyGreenTable/useLeafyGreenTable.styles.ts
@@ -2,7 +2,6 @@ import { css } from '@leafygreen-ui/emotion';
 import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 
-// Unselected checkboxes
 export const disabledTableRowCheckStyles: Record<Theme, string> = {
   [Theme.Light]: css`
     input + div {
